add csv to json in Utils class
